refactor: split utils/data.py into three focused modules (Issue 6)
