+++ python/agActor/Commands/AgCmd.py
@@ -11,7 +11,7 @@ from pfs.utils.database.gaia import GaiaDB
 from agActor import field_acquisition
 from agActor.Controllers.ag import ag
 from agActor.catalog import pfs_design
-from agActor.utils import actorCalls, data as data_utils
+from agActor.utils import actorCalls, queries as data_utils
 from agActor.utils.focus import focus
 from agActor.utils.actorCalls import send_guide_offsets
 from agActor.utils.telescope_center import telCenter as tel_center
